refactor: Service Layer 삭제
