+++ src/main/java/Main.java
@@ -1,6 +1,6 @@
 import controller.BlockController;
 import java.util.Scanner;
-import service.BlockService;
+import model.LadderMaker;
 import strategy.RandomPassGenerator;
 import view.InputView;
 import view.OutputView;
@@ -10,8 +10,8 @@ public class Main {
     public static void main(String[] args) {
         InputView inputView = new InputView(new Scanner(System.in));
         OutputView outputView = new OutputView();
-        BlockService blockService = new BlockService(new RandomPassGenerator());
-        BlockController blockController = new BlockController(inputView, outputView, blockService);
+        LadderMaker ladderMaker = new LadderMaker(new RandomPassGenerator());
+        BlockController blockController = new BlockController(inputView, outputView, ladderMaker);
 
         blockController.run();
     }
